Fix "Restarting server due to time wrapping"
sv.time check as well and early warning system

need to actually shutdown with an svs.time style wrap

+++ src/server/sv_ccmds.cpp
@@ -653,7 +653,7 @@ static void SV_Status_f( void )
 			state,
 			displayName,
 			spaces,
-			svs.time - cl->lastPacketTime,
+			(int)(svs.time - cl->lastPacketTime), // do we prefer a different modifier here for the impossible case where its outside int?
 			s,
 			cl->netchan.qport,
 			SV_ClientRate(cl)
